requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Database & ORM
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.5.0

# System Monitoring
psutil==5.9.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2